Simplify gradual font size (molten leading)
`min()`, `max()` and `clamp()` are now available in major browsers
(since Edge 79, released 15 Jan 2020; FF 75, released 8 Apr 2020; and Chrome 79 released 10 Dec 2019)
so we can use it instead of a media query to make the font size adapt to the viewport size.

Using `clamp()` also allows us to introduce a maximum font size as well,
as was requested in #47.